shorten the regex in an expect_error() test to avoid error by "covr"
The last test in test pc20slMatrix.r was giving error when running the
report functions from package "covr", despite the fact that `R CMD check`
and `devtools::test()` were ok. I cut the regex just before the
first single quote. This resolved the issue. However before that I tried
to change the single quote with a dot but that didn't work for 'covr'.

This should resolve the mistery that Coveralls wasn't getting updates
from successful build of 'lagged' on TravisCI.